Updated printf implementation
Current iteration of printf implementation. Uses heap memory to store format specifier data and is very slow (comparatively to stdio printf) as a result. Implemented a format specifier pre-checker that controls what format specifier data is passed to the rest of the function to be printed. No working float print.